requirements.txt
pandas
numpy
scikit-learn
requests
pyarrow
streamlit
python-dotenv